requirements.txt
streamlit
PyPDF2
faiss-cpu
sentence-transformers
transformers
torch
ibm-watsonx-ai
